test(state-mgmt): add missing coverage for EnsureStatePath, relocate misplaced Python version listing func
